FitnessController 생성, IndexController 역할 분리, html 파일 역할 분리
