+++ src/main/java/net/sesoc/web1/controller/IndexController.java
@@ -1,0 +1,12 @@
+package net.sesoc.web1.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class IndexController {
+	@GetMapping
+	public String index() {
+		return "index";
+	}
+}
