+++ src/main/resources/templates/join.html
@@ -1,0 +1,19 @@
+<!DOCTYPE html>
+<html xmlns:th="http://www.thymeleaf.org"> 
+<head>
+  <meta charset="UTF-8">
+  <title>Insert title here</title>
+</head>
+<body>
+  <h2>[ 회원 가입 ] </h2>
+  <form th:action="@{/join}" method="post">
+    <label> 아이디 : 
+      <input type="text" id="usrid" name="usrid" placeholder="회원아이디">
+    </label><br>
+    <label> 비밀번호 : 
+      <input type="password" id="usrpwd" name="usrpwd" placeholder="회원비밀번호">
+    </label><br>
+    <input type="submit" value="회원가입"> <input type="reset" value="취소"> 
+  </form>
+</body>
+</html>
